+++ packages/snippets/src/acro-dynamic-form/index.ts
@@ -1,15 +1,16 @@
 import AcroDynamicForm from './src/index.vue';
-import { defineAcroDynamicFormFields } from './src/utils/defineAcroDynamicFormFields';
+import { defineAcroDynamicFormField, defineAcroDynamicFormFields } from './src/utils/defineAcroDynamicFormFields';
 import type { AcroDynamicFormField, AcroDynamicFormProps } from './src/types';
 
 type AcroDynamicFormInstance = InstanceType<typeof AcroDynamicForm>;
 
 export {
   AcroDynamicForm,
-  AcroDynamicFormProps,
+
   AcroDynamicFormInstance,
-
-  defineAcroDynamicFormFields,
-
+  AcroDynamicFormProps,
   AcroDynamicFormField,
+
+  defineAcroDynamicFormField,
+  defineAcroDynamicFormFields,
 };