+++ packages/snippets/src/acro-dynamic-form/src/utils/defineAcroDynamicFormFields.ts
@@ -1,6 +1,13 @@
 import type { AcroDynamicFormField } from '../types';
 
 /**
+ * 定义单个字段配置
+ */
+export function defineAcroDynamicFormField(field: AcroDynamicFormField) {
+  return field;
+}
+
+/**
  * 定义字段配置列表
  */
 export function defineAcroDynamicFormFields(fields: AcroDynamicFormField[]) {